pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["packages*"]

[project]
name = "autoresearch-robotics"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "pydantic>=2.0",
  "supabase>=2.4",
  "google-genai>=1.0.0",
  "google-api-python-client>=2.120",
  "yt-dlp>=2024.4",
  "modal>=0.62",
  "torch>=2.2",
  "torch-geometric>=2.5",
  "mujoco>=3.1",
  "numpy>=1.26",
  "pink>=3.1",
  "pinocchio>=3.1",
  "python-dotenv>=1.0",
  "httpx>=0.27",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "httpx>=0.27"]

[tool.pytest.ini_options]
testpaths = ["tests"]
